external/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
poe-api==0.3.1
PySocks==1.7.1
python-socks==2.3.0
requests==2.30.0
simple-websocket==0.10.0
urllib3==1.26.6
websocket-client==1.5.1
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0
